refactor(SetTokenDialog): refactor to TS, modularize config logic into separate components
