+++ client/src/components/Input/SetTokenDialog/index.ts
@@ -1,0 +1,1 @@
+export { default as SetTokenDialog } from './SetTokenDialog';
