+++ client/src/components/Input/SubmitButton.jsx
@@ -1,8 +1,8 @@
 import React, { useState } from 'react';
-import StopGeneratingIcon from '../svg/StopGeneratingIcon';
+import { StopGeneratingIcon } from '~/components';
 import { Settings } from 'lucide-react';
-import SetTokenDialog from './SetTokenDialog';
-import store from '../../store';
+import { SetTokenDialog } from './SetTokenDialog';
+import store from '~/store';
 
 export default function SubmitButton({
   endpoint,
