+++ client/src/components/svg/index.ts
@@ -4,3 +4,4 @@ export { default as BingIcon } from './BingIcon';
 export { default as CogIcon } from './CogIcon';
 export { default as Spinner } from './Spinner';
 export { default as MessagesSquared } from './MessagesSquared';
+export { default as StopGeneratingIcon } from './StopGeneratingIcon';
